+++ apps/client/src/app/(auth)/layout.tsx
@@ -70,10 +70,9 @@ export default function AuthLayout({ children }: { children: ReactNode }) {
         try {
           const response = await axios.get(endPoints.userDetails)
           const userId = response.data.user._id
-          const { email, privateKey, role, selectedWorkspaceId } = response.data.user
+          const { email, privateKey, role, selectedWorkspaceId, reduceCarbonEmissions } = response.data.user
           const { name: selectedWorkspaceName, clientId, clientSecret } = response.data.workspace
           const hasActiveSubscription = response.data.hasActiveSubscription
-          const { useEnergySaver, useOptimizedAPICalls } = response.data.sustainabilitySettings
 
           if (response.data.subscription) {
             const { selectedPlan, createdAt, expiresAt, remainingCredits } = response.data.subscription
@@ -86,7 +85,7 @@ export default function AuthLayout({ children }: { children: ReactNode }) {
 
           localStorage.setItem("clientId", clientId)
           localStorage.setItem("clientSecret", clientSecret)
-          dispatch("setUserState", { userId, email, privateKey, role, selectedWorkspaceId, selectedWorkspaceName, clientId, clientSecret, hasActiveSubscription, useEnergySaver, useOptimizedAPICalls })
+          dispatch("setUserState", { userId, email, privateKey, role, selectedWorkspaceId, selectedWorkspaceName, clientId, clientSecret, hasActiveSubscription, reduceCarbonEmissions })
           dispatch("setUserState", { isAuthorized: true })
           setAuthorized(true)
         }
@@ -121,7 +120,7 @@ export default function AuthLayout({ children }: { children: ReactNode }) {
 
   return (
     <Suspense condition={!isLoading} fallback={<Loading />}>
-      <Suspense condition={!isAuthorized} fallback={<div style={userState.useEnergySaver ? { filter: "brightness(60%) contrast(80%)" } : {}}>{children}</div>}>
+      <Suspense condition={!isAuthorized} fallback={children}>
         <Suspense condition={authStep === 1} fallback={null}>
           <div className="container-center">
             <form className="box" onSubmit={generatePassKey}>