+++ apps/server/src/api/api.module.ts
@@ -4,9 +4,8 @@ import { SubscriptionModule } from "./subscription/subscription.module"
 import { ProductsModule } from "./products/products.module"
 import { ApiReferenceModule } from "./apireference/apireference.module"
 import { WorkspaceModule } from "./workspace/workspace.module"
-import { SustainabilityModule } from "./sustainability/sustainability.module"
 
 @Module({
-  imports: [UserModule, SubscriptionModule, ProductsModule, ApiReferenceModule, WorkspaceModule, SustainabilityModule]
+  imports: [UserModule, SubscriptionModule, ProductsModule, ApiReferenceModule, WorkspaceModule]
 })
 export class ApiModule { }