+++ apps/server/src/api/sustainability/commands/update-settings.command.ts
@@ -1,6 +1,0 @@
-import { SustainabilityModel } from "../models/sustainability.model"
-
-export default async function updateSustainabilitySettingsCommand(userId: string, useEnergySaver: boolean, useOptimizedAPICalls: boolean) {
-  await SustainabilityModel.findOneAndUpdate({ userId }, { useEnergySaver, useOptimizedAPICalls })
-  return true
-}