+++ apps/server/src/api/sustainability/dto/update-sustainability.dto.ts
@@ -1,6 +1,0 @@
-import { PartialType } from "@nestjs/mapped-types"
-
-export class UpdateSustainabilityDto {
-  readonly useEnergySaver: boolean
-  readonly useOptimizedAPICalls: boolean
-}