+++ apps/server/src/api/sustainability/queries/fetch-sustainability-settings.query.ts
@@ -1,5 +1,0 @@
-import { SustainabilityModel } from "../models/sustainability.model"
-
-export default async function fetchSustainabilitySettings(userId: string) {
-  return await SustainabilityModel.findOne({ userId })
-}