+++ apps/server/src/api/user/commands/update-carbon-settings.command.ts
@@ -1,0 +1,6 @@
+import { UserModel } from "../models/user.model"
+
+export default async function updateCarbonSettings(userId: string, value: boolean) {
+  await UserModel.findByIdAndUpdate(userId, { reduceCarbonEmissions: value })
+  return true
+}